android communication

+++ ConsoleApp1/OBJGenerator.cs
@@ -2,10 +2,12 @@
 using Newtonsoft.Json.Linq;
 using System.Diagnostics;
 using IGtoOBJGen;
+using SharpAdbClient;
 using System.Net.Mail;
 using System.Reflection.Emit;
 using System.Xml.Linq;
 using System.Text.RegularExpressions;
+using System.Net.Http.Headers;
 
 /*
 
@@ -14,13 +16,26 @@ What's going on in this file? It's just the main file. Why is it so messy? Good 
  */
 class OBJGenerator
 {
-    static void Main(string[] args)
+     static void Main(string[] args)
     {
         StreamReader file;
         string eventName;
         string strPath = Environment.GetFolderPath(Environment.SpecialFolder.DesktopDirectory);
-        var watch = new Stopwatch(); 
+        var watch = new Stopwatch();
         watch.Start();
+        
+        /*
+        AdbServer server = new AdbServer();
+        var result = server.StartServer(@"C:\Users\uclav\AppData\Local\Android\Sdk\platform-tools\adb.exe", restartServerIfNewer: false);
+        AdbClient client = new AdbClient();
+        var devices = client.GetDevices();
+        foreach (var device in devices)
+        {
+            Console.WriteLine(device.Model);
+        }*/
+
+        IGTracks trackHandler = new IGTracks(); 
+        IGBoxes boxHandler = new IGBoxes();
 
         if (args.Length == 0){
             file = File.OpenText($"C:\\Users\\uclav\\Source\\Repos\\jdmalham\\IG-File-OBJ-Generator\\ConsoleApp1\\IGdata\\Event_1096322990");
@@ -42,9 +57,9 @@ class OBJGenerator
             file = File.OpenText($"{args[0]}.tmp");
         }
 
-        //Read in IG file as JSON Object. Thanks Newtonsoft
         string user = Environment.GetFolderPath(Environment.SpecialFolder.UserProfile);
-        
+
+        //Read in IG file as JSON Object. Thanks Newtonsoft
         JsonTextReader reader = new JsonTextReader(file);
         JObject o2 = (JObject)JToken.ReadFrom(reader);
         
@@ -78,20 +93,27 @@ class OBJGenerator
         }
 
         //Get the data that is needed to generate track objects from the IG file (look at definition for explanation of the name)
+        {
+            List<TrackExtrasData> listicle = trackHandler.trackExtrasParse(o2);
 
-        List<TrackExtrasData> listicle = IGTracks.trackExtrasParse(o2);
+            trackHandler.trackCubicBezierCurve(listicle, 32, eventName);  //Create the cubic bezier curve object file based off the track data
 
-        IGTracks.trackCubicBezierCurve(listicle,32,eventName);  //Create the cubic bezier curve object file based off the track data
+            var n = trackHandler.photonParse(o2); // Get photon data
+            trackHandler.generatePhotonModels(n, eventName); // This one is a mystery. I don't know what it does. I can't figure it out. WHAT DOES GENERATE PHOTON MODELS MEAN?????
 
-        var n = IGTracks.photonParse(o2); // Get photon data
-        IGTracks.generatePhotonModels(n,eventName); // This one is a mystery. I don't know what it does. I can't figure it out. WHAT DOES GENERATE PHOTON MODELS MEAN?????
-        
-        List<MuonChamberData> list = IGBoxes.muonChamberParse(o2); // Get muon chamber data
-        IGBoxes.generateMuonChamberModels(list); // Generate muon chamber obj files
-
+            List<MuonChamberData> list = IGBoxes.muonChamberParse(o2); // Get muon chamber data
+            IGBoxes.generateMuonChamberModels(list); // Generate muon chamber obj files
+        }
         file.Close();
-        File.Delete($"{args[0]}.tmp");
+        
+        if (args.Length > 0)
+        {
+            File.Delete($"{args[0]}.tmp");
+        }
 
+        Communicate bridge = new Communicate(@"C:\Users\uclav\AppData\Local\Android\Sdk\platform-tools\adb.exe");
+        bridge.DownloadFiles("tracks.obj");
+        bridge.UploadFiles(trackHandler.filePaths);        
         Console.WriteLine($"Total Execution Time: {watch.ElapsedMilliseconds} ms"); // See how fast code runs. Code goes brrrrrrr on fancy office pc. It makes me happy. :)
     }
 }
